requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
SQLAlchemy==2.0.27
aiosqlite==0.19.0
pydantic==2.6.1